Tried to refactor
