+++ lib/features/dashboard/presentation/widget/all_supplications.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+
 class AllDuas extends StatefulWidget {
   const AllDuas({Key? key}) : super(key: key);
 
@@ -6,11 +7,10 @@ class AllDuas extends StatefulWidget {
   State<AllDuas> createState() => _AllDuasState();
 }
 
-
-
 class _AllDuasState extends State<AllDuas> {
   final List<String> entries = <String>['A', 'B', 'C'];
   final List<int> colorCodes = <int>[600, 500, 100];
+
   @override
   Widget build(BuildContext context) {
     return ListView.builder(
@@ -18,14 +18,12 @@ class _AllDuasState extends State<AllDuas> {
       itemCount: entries.length,
       itemBuilder: (BuildContext context, int index) {
         return Container(
-          height: 70,
-          color: Colors.green[colorCodes[index]],
-          child: Center(child: (Text('Dua ${entries[index]}')),)
-        );
+            height: 70,
+            color: Colors.green[colorCodes[index]],
+            child: Center(
+              child: (Text('Dua ${entries[index]}')),
+            ));
       },
-      
     );
   }
-
-
 }
