--- lib/text.txt
@@ -1,19 +1,0 @@
- ServiceCard(
-            containerWidth: 123,
-            containerHeight: 50,
-            nameImage: 'taxi.jpg',
-            nameEng: 'Taxi'
-        ),
-        ServiceCard(
-            containerWidth: 123,
-            containerHeight: 50,
-            nameImage: 'taxi.jpg',
-            nameEng: 'Taxi'
-        ),
-        ServiceCard(
-            containerWidth: 123,
-            containerHeight: 50,
-            nameImage: 'taxi.jpg',
-            nameEng: 'Taxi'
-        ),
-          ),
